Chore: Rename class and namespaces of purchases' class
