+++ supermarket_sales_manegement/UserControls/Purchase/PurchaseUserControl.resx
